Fix some 64-bit build warnings.
--HG--
branch : R2_14

+++ Src/Common/UniFile.cpp
@@ -425,9 +425,9 @@ bool UniMemFile::ReadString(String & line, bool * lossy)
 static void Append(String &strBuffer, LPCTSTR pchTail,
 		int cchTail, int cchBufferMin)
 {
-	int cchBuffer = strBuffer.capacity();
-	int cchHead = strBuffer.length();
-	int cchLength = cchHead + cchTail;
+	size_t cchBuffer = strBuffer.capacity();
+	size_t cchHead = strBuffer.length();
+	size_t cchLength = cchHead + cchTail;
 	while (cchBuffer < cchLength)
 	{
 		ASSERT((cchBufferMin & cchBufferMin - 1) == 0); // must be a power of 2
@@ -817,7 +817,7 @@ bool UniStdioFile::ReadBom()
 	if (buff == NULL)
 		return false;
 
-	int bytes = fread(buff, 1, max_size, m_fp);
+	size_t bytes = fread(buff, 1, max_size, m_fp);
 	m_data = 0;
 	m_charsize = 1;
 	bool unicode = false;
@@ -919,8 +919,8 @@ bool UniStdioFile::WriteString(const String & line)
 	// shortcut the easy cases
 	if (m_unicoding == ucr::UCS2LE)
 	{
-		unsigned int bytes = line.length() * sizeof(TCHAR);
-		unsigned int wbytes = fwrite(line.c_str(), 1, bytes, m_fp);
+		size_t bytes = line.length() * sizeof(TCHAR);
+		size_t wbytes = fwrite(line.c_str(), 1, bytes, m_fp);
 		if (wbytes != bytes)
 			return false;
 		return true;
@@ -931,10 +931,10 @@ bool UniStdioFile::WriteString(const String & line)
 	int codepage1 = 0;
 	ucr::getInternalEncoding(&unicoding1, &codepage1); // What String & TCHARs represent
 	const unsigned char * src = (const UCHAR *)line.c_str();
-	int srcbytes = line.length() * sizeof(TCHAR);
+	size_t srcbytes = line.length() * sizeof(TCHAR);
 	bool lossy = ucr::convert(unicoding1, codepage1, src, srcbytes, (ucr::UNICODESET)m_unicoding, m_codepage, buff);
 	// TODO: What to do about lossy conversion ?
-	unsigned int wbytes = fwrite(buff->ptr, 1, buff->size, m_fp);
+	size_t wbytes = fwrite(buff->ptr, 1, buff->size, m_fp);
 	if (wbytes != buff->size)
 		return false;
 	return true;
